+++ src/backend_starter/generator.py
@@ -6,6 +6,8 @@ from backend_starter.templates import(
     CHANGELOG_TEMPLATE,
     GITIGNORE_TEMPLATE,
     README_TEMPLATE,
+    AGENT_TEMPLATE,
+    PROJECT_TEMPLATE,
 )
 
 
@@ -41,6 +43,16 @@ def create_project(project_name: str, create_venv: bool = False) -> None:
 
     (project_path / ".gitignore").write_text(
         GITIGNORE_TEMPLATE,
+        encoding="utf-8",
+    )
+
+    (project_path / "AGENTS.md").write_text(
+        AGENT_TEMPLATE,
+        encoding="utf-8",
+    )
+
+    (project_path / "PROJECT.md").write_text(
+        PROJECT_TEMPLATE,
         encoding="utf-8",
     )